Move getSlotLabwareName to shared util
