+++ app/src/organisms/Devices/ProtocolRun/SetupLiquids/__tests__/utils.test.ts
@@ -2,12 +2,10 @@ import {
   getWellFillFromLabwareId,
   getTotalVolumePerLiquidId,
   getTotalVolumePerLiquidLabwarePair,
-  getSlotLabwareName,
   getLiquidsByIdForLabware,
   getWellGroupForLiquidId,
   getWellRangeForLiquidLabwarePair,
 } from '../utils'
-import { getLabwareDisplayName } from '@opentrons/shared-data'
 import type { LabwareByLiquidId, Liquid } from '@opentrons/api-client'
 
 const LABWARE_ID =
@@ -108,49 +106,6 @@ const MOCK_LABWARE_BY_LIQUID_ID: LabwareByLiquidId = {
   ],
 }
 
-const MOCK_LOAD_LABWARE_COMMANDS = [
-  {
-    commandType: 'loadLabware',
-    params: {
-      location: {
-        slotName: '5',
-      },
-    },
-    result: {
-      labwareId:
-        '60e8b050-3412-11eb-ad93-ed232a2337cf:opentrons/corning_24_wellplate_3.4ml_flat/1',
-      definition: {},
-    },
-  },
-]
-
-const MOCK_COMMANDS = [
-  {
-    commandType: 'loadLabware',
-    params: {
-      location: {
-        moduleId: '12345',
-      },
-    },
-    result: {
-      labwareId:
-        '60e8b050-3412-11eb-ad93-ed232a2337cf:opentrons/corning_24_wellplate_3.4ml_flat/1',
-      definition: {},
-    },
-  },
-  {
-    commandType: 'loadModule',
-    params: {
-      location: {
-        slotName: '4',
-      },
-    },
-    result: {
-      moduleId: '12345',
-    },
-  },
-]
-
 const MOCK_LABWARE_BY_LIQUID_ID_FOR_LABWARE = {
   '4': [
     {
@@ -223,11 +178,6 @@ const MOCK_VOLUME_BY_WELL = {
   D4: 100,
 }
 
-jest.mock('@opentrons/shared-data')
-const mockGetLabwareDisplayName = getLabwareDisplayName as jest.MockedFunction<
-  typeof getLabwareDisplayName
->
-
 describe('getWellFillFromLabwareId', () => {
   it('returns wellfill object for the labwareId', () => {
     const expected = {
@@ -285,32 +235,6 @@ describe('getTotalVolumePerLiquidLabwarePair', () => {
         MOCK_LABWARE_BY_LIQUID_ID
       )
     ).toEqual(expected)
-  })
-})
-
-describe('getSlotLabwareName', () => {
-  beforeEach(() => {
-    mockGetLabwareDisplayName.mockReturnValue(
-      'Corning 24 Well Plate 3.4 mL Flat'
-    )
-  })
-  it('returns labware name and slot number for labware id', () => {
-    const expected = {
-      slotName: '5',
-      labwareName: 'Corning 24 Well Plate 3.4 mL Flat',
-    }
-    expect(
-      getSlotLabwareName(LABWARE_ID, MOCK_LOAD_LABWARE_COMMANDS as any)
-    ).toEqual(expected)
-  })
-  it('returns the module slot number if the labware is on a module', () => {
-    const expected = {
-      slotName: '4',
-      labwareName: 'Corning 24 Well Plate 3.4 mL Flat',
-    }
-    expect(getSlotLabwareName(LABWARE_ID, MOCK_COMMANDS as any)).toEqual(
-      expected
-    )
   })
 })
