+++ app/src/organisms/Devices/ProtocolRun/StepText.tsx
@@ -4,7 +4,7 @@ import { Flex, ALIGN_CENTER, SPACING, TYPOGRAPHY } from '@opentrons/components'
 import { getLabwareDisplayName } from '@opentrons/shared-data'
 import { StyledText } from '../../../atoms/text'
 import { getLabwareLocation } from '../ProtocolRun/utils/getLabwareLocation'
-import { getSlotLabwareName } from './SetupLiquids/utils'
+import { getSlotLabwareName } from './utils/getSlotLabwareName'
 
 import {
   useLabwareRenderInfoForRunById,
